[Condition] Collect deduped field paths in parsing
In some scenarios the condition evaluation can require building a picture of fields the condition depends on.
To that end, it is useful to build the definitive list of unique field paths as part of parsing to avoid consumers
having to re-iterate over the condition, likely iterating over the same criteria multiple times because of normalisation.