Breaking: Refactor ApiClient
Instead of creating an Axios instance for every request, allow for an Axios instance to be shared across all requests for a given client.
This is required to enable the use of caching interceptors.

As a consequence, interceptors can no longer be used for request-centric logic (abort/auth) and instead a decorator must be used.