Set `id_token_hint` in authorize URL from session
When an ID token is present in an existing session (ie. re-authentication for a session with expired access/refresh tokens), pass the ID token to the authorisation server via the `id_token_hint` request parameter.
While this may not be supported by all OIDC providers, the ones which do support it can use it to validate the user being re-authenticated is the same as the current SSO session, if one exists.

See: https://openid.net/specs/openid-connect-core-1_0.html#AuthRequest