Condition OpenID RP-initiated logout
By default, only initiate an RP-initiated logout if the ID token is present in the session.
This is because providing the logout `id_token_hint` is recommended and its absence could result in unexpected behaviour.
As an alternative, when no ID token is present in the session (ie. the session likely expired and is empty), then simply destroy the session and call the next middleware in the chain.

RP-initiated logout can be forced using the `forceRpInitiatedLogout` option.